supplier_finder_deep_research_agent/requirements.txt
google-adk==1.19.0
google-cloud-aiplatform==1.128.0
google-genai==1.52.0
pydantic==2.12.5
requests
python-dotenv
google-auth
google-cloud-aiplatform[adk,agent_engines]>=1.38.0
